Update solvent module layout
